requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
passlib==1.7.4
pip==25.0.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
